compiler: mk sync with goken!

+++ compilers/5c/cgen.c
@@ -380,7 +380,7 @@ cgenrel(Node *n, Node *nn, bool inrel)
                     diag(n, "DOT and no offset");
                     break;
                 }
-                nod.xoffset += (long)r->vconst;
+                nod.xoffset += (int32)r->vconst;
                 nod.type = n->type;
                 cgen(&nod, nn);
             }
@@ -1050,7 +1050,7 @@ sugen(Node *n, Node *nn, long w)
                 diag(n, "DOT and no offset");
                 break;
             }
-            nod1.xoffset += (long)r->vconst;
+            nod1.xoffset += (int32)r->vconst;
             nod1.type = n->type;
             sugen(&nod1, nn, w);
         }
