+++ compilers/8c/cgen.c
@@ -1030,7 +1030,7 @@ cgen(Node *n, Node *nn)
             diag(n, "DOT and no offset");
             break;
         }
-        nod.xoffset += (long)r->vconst;
+        nod.xoffset += (int32)r->vconst;
         nod.type = n->type;
         cgen(&nod, nn);
         break;
@@ -1512,7 +1512,7 @@ sugen(Node *n, Node *nn, long w)
             diag(n, "DOT and no offset");
             break;
         }
-        nod1.xoffset += (long)r->vconst;
+        nod1.xoffset += (int32)r->vconst;
         nod1.type = n->type;
         sugen(&nod1, nn, w);
         break;
