+++ compilers/8c/gc.h
@@ -1,8 +1,8 @@
 /*s: 8c/gc.h */
 #include	"../cc/cc.h"
 
-#include	<common.out.h>
-#include	<8.out.h>
+#include	<obj/common.out.h>
+#include	<obj/8.out.h>
 
 /*
  * 8c/386
@@ -60,7 +60,8 @@ extern struct Idx idx;
 /*s: struct [[Adr]](x86) */
 struct	Adr
 {
-    long	offset;
+    // claude: int32, not long, like in the kencc lineage (see changes.txt)
+    int32	offset;
     double	dval;
     char	sval[NSNAME];
     Sym*	sym;
@@ -130,7 +131,7 @@ struct	Var
     char	symkind;
     //enum<type_kind>
     char	etype;
-    long	offset;
+    int32	offset;
 };
 /*e: struct [[Var]] */
