+++ compilers/cc2/pswt.c
@@ -57,7 +57,10 @@ doswit(Node *n)
         if(isv)
             q->val = c->val;
         else
-            q->val = (long)c->val;	/* cast ensures correct value for 32-bit switch on 64-bit architecture */
+            // claude: int32, not long: a 64-bit long keeps hex constants
+            // like 0xC0A1E5CE positive, changing the switch sort order;
+            // see tests/c/misc/switch_hex.c
+            q->val = (int32)c->val;	/* cast ensures correct value for 32-bit switch on 64-bit architecture */
         q++;
     }
     qsort(iq, nc, sizeof(C1), swcmp);
@@ -113,7 +116,7 @@ doswit(Node *n)
         for(j = i; j < nc; j++){
             if((iq[j].val>>32) != iqh[nh].val)
                 break;
-            q->val = (long)iq[j].val;
+            q->val = (int32)iq[j].val;
             q->label = iq[j].label;
             q++;
         }
@@ -211,7 +214,7 @@ ieeedtod(Ieee *ieee, double native)
     fr = modf(fr*f, &ho);
     ieee->l = ho;
     ieee->l <<= 16;
-    ieee->l |= (long)(fr*f);
+    ieee->l |= (int32)(fr*f);
 }
 /*e: function [[ieeedtod]] */
 /*e: cc2/pswt.c */
